last update 2018.06.28
ヽ(ﾟ∀｡)ﾉ

+++ NyaEffect.h
@@ -17,8 +17,8 @@ namespace H2NLIB
 	{
 	public:
 		int interval_time_frame_;		//!< アニメーションインターバル時間(フレーム指定)
-		int grid_x_;					//!< 描画x座標
-		int grid_y_;					//!< 描画y座標
+		double grid_x_;					//!< 描画x座標
+		double grid_y_;					//!< 描画y座標
 	};
 
 	class EffectPropertyX2
@@ -27,8 +27,8 @@ namespace H2NLIB
 		double draw_move_x_;
 		double draw_move_y_;
 		int interval_time_frame_;
-		int grid_x_;
-		int grid_y_;
+		double grid_x_;
+		double grid_y_;
 	};
 
 	class EffectAnimation1
